Create Form component and custom hook
